Actually Added tooltips
This works for ingredients, instructions, and tools. Check out the
asparagus soup example for syntax when writing new recipes. The
definition of what goes in a tooltip is recipe-by-recipe. No universal
list. WEll, there is a universal picture list. but that's it.

Also, tooltips appear whereever your mouse cursor is, and the box will
only spill off screen if there's room. so when the application fills up
a window the position of the box is always sensible. Try it out.