Fix issue where recipe step buttons aren't updated
When going from the search page to the recipe page, the list of coloured
step buttons weren't updated until you scroll for the first time. Now
the list is updated once the animation over to the recipe page is over.

It's not perfect, but it's an improvement. Feel free to change to
something better if you want, but I think this is fine.